Trying to fix the move code once more

+++ Team22258/src/main/java/org/firstinspires/ftc/team22258/teleop/opmode_TeleOp.java
@@ -120,7 +120,7 @@ public class opmode_TeleOp extends LinearOpMode {
     }
     if (true) {
       // Power Control
-      MotorPowerNormalizer = ( Math.max( Math.max( (double) Math.abs(gamepad1.left_stick_x), (double) Math.abs(gamepad1.left_stick_y) ), (double) Math.abs(gamepad1.right_stick_x) ) / ( Math.max( Math.max( (double) Math.abs(FLMotorPower), (double) Math.abs(FRMotorPower) ), Math.max( (double) Math.abs(BLMotorPower), (double) Math.abs(BRMotorPower) ) ) ) );
+      MotorPowerNormalizer =  (double) ( Math.max( Math.max( Math.abs(gamepad1.left_stick_x), Math.abs(gamepad1.left_stick_y) ), Math.abs(gamepad1.right_stick_x) ) / ( Math.max( Math.max( Math.abs(FLMotorPower), Math.abs(FRMotorPower) ), Math.max( Math.abs(BLMotorPower), Math.abs(BRMotorPower) ) ) ) );
       FLMotorPower = (int) (FLMotorPower * MotorPowerNormalizer);
       FRMotorPower = (int) (FRMotorPower * MotorPowerNormalizer);
       BLMotorPower = (int) (BLMotorPower * MotorPowerNormalizer);
